Remove TileAndDecomposeAttention
