+++ compiler/src/iree/compiler/Dialect/LinalgExt/Transforms/CMakeLists.txt
@@ -35,7 +35,6 @@ iree_cc_library(
     "PassDetail.h"
     "Passes.cpp"
     "SplitReduction.cpp"
-    "TileAttention.cpp"
     "TilingInterfaceImpl.cpp"
   DEPS
     ::PassesIncGen
