+++ compiler/src/iree/compiler/Dialect/LinalgExt/Transforms/Passes.h
@@ -46,26 +46,14 @@ createDecomposeWinogradTransformPass();
 std::unique_ptr<InterfacePass<mlir::FunctionOpInterface>>
 createConvertConv2DToWinogradPass();
 
-IREE::LinalgExt::AttentionOp
-tileAttention(IREE::LinalgExt::AttentionOp attnOp,
-              SmallVectorImpl<Operation *> &ops, RewriterBase &rewriter,
-              std::optional<uint64_t> tileSize = std::nullopt);
-
-void decomposeTiledAttention(IREE::LinalgExt::AttentionOp tiledAttnOp,
-                             SmallVectorImpl<Operation *> &ops,
-                             RewriterBase &rewriter,
-                             std::optional<uint64_t> tileSize = std::nullopt);
-
 void convertToOnlineAttention(IREE::LinalgExt::AttentionOp attnOp,
                               SmallVectorImpl<Operation *> &ops,
                               RewriterBase &rewriter);
 
-// Creates a pass to tile the attention op along the reduction dim.
-std::unique_ptr<Pass> createTileAttentionPass();
-
 // Creates a pass to convert the attention op into a sequence of linalg ops.
 std::unique_ptr<Pass> createDecomposeAttentionPass();
 
+// Creates a pass to convert attention op into online_attention op.
 std::unique_ptr<Pass> createConvertAttentionToOnlineAttentionPass();
 
 //===---------------------------------------------------------------------===//
