+++ compiler/src/iree/compiler/Dialect/LinalgExt/Transforms/test/BUILD.bazel
@@ -18,13 +18,11 @@ iree_lit_test_suite(
         [
             "conv2d_to_winograd.mlir",
             "convert_to_loops.mlir",
-            "decompose_attention.mlir",
             "decompose_online_attention.mlir",
             "decompose_winograd.mlir",
             "distribution.mlir",
             "pad_contraction_to_block_size.mlir",
             "split_reduction.mlir",
-            "tile_attention.mlir",
             "tiling.mlir",
         ],
         include = ["*.mlir"],
