+++ compiler/src/iree/compiler/Dialect/LinalgExt/Transforms/test/CMakeLists.txt
@@ -16,13 +16,11 @@ iree_lit_test_suite(
   SRCS
     "conv2d_to_winograd.mlir"
     "convert_to_loops.mlir"
-    "decompose_attention.mlir"
     "decompose_online_attention.mlir"
     "decompose_winograd.mlir"
     "distribution.mlir"
     "pad_contraction_to_block_size.mlir"
     "split_reduction.mlir"
-    "tile_attention.mlir"
     "tiling.mlir"
   TOOLS
     FileCheck
